WIP Important moment as we would with CLI
This is not working yet. I don't even have the same syntax exactly yet - is close, but not there.

Current failure is because I'm unable so far to persuade Closure to let me import a CommonJS module from a ES2015 import statement. And this is necessary in turn, because TypeScript can't emit the right code to expose code that can be imported to the application using:

import * as foo from 'foo'

... When emitting a 2015 module.

Help wanted. :-)